requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1
openai>=1.30.1
langchain>=0.1.20